+++ astropy/units/deprecated.py
@@ -34,6 +34,11 @@ def _initialize_module():
 
     # add only some *prefixes* as deprecated units
     _add_prefixes(astrophys.solMass, namespace=_ns, prefixes=True)
+    _add_prefixes(astrophys.jupiterMass, namespace=_ns, prefixes=True)
+    _add_prefixes(astrophys.earthMass, namespace=_ns, prefixes=True)
+    _add_prefixes(astrophys.solRad, namespace=_ns, prefixes=True)
+    _add_prefixes(astrophys.jupiterRad, namespace=_ns, prefixes=True)
+    _add_prefixes(astrophys.earthRad, namespace=_ns, prefixes=True)
 
 
 _initialize_module()